+++ cli/tests/test_log_command.rs
@@ -45,6 +45,7 @@ fn test_log_with_no_template() {
 
     For more information, try '--help'.
     Hint: The following template aliases are defined:
+    - builtin_commit_summary_redacted
     - builtin_config_list
     - builtin_config_list_detailed
     - builtin_draft_commit_description