+++ cli/tests/test_show_command.rs
@@ -334,6 +334,7 @@ fn test_show_with_no_template() {
 
     For more information, try '--help'.
     Hint: The following template aliases are defined:
+    - builtin_commit_summary_redacted
     - builtin_config_list
     - builtin_config_list_detailed
     - builtin_draft_commit_description